requirements.txt
Django==3.2.20
gunicorn==20.1.0
psycopg2==2.9.5
Pyrebase4==4.7.1